Introduce the concept of a runtime record-type descriptor.
In previous iterations of this proposal, the predicate procedure
of a record type served a second purpose, namely that of a runtime
representaton of the record type descriptor to be used in the
procedural interface. This led to some misunderstandings and a
couple of questions raised in
http://srfi-email.schemers.org/srfi-136/msg/4000979. These
questions have been addressed in this modification of the proposal,
which introduces the notion of a runtime record-type descriptor, which
can be extracted from the (syntactic) record-type descriptor and
which frees the record-type predicate from having to serve two purposes.